+++ CHANGELOG.md
@@ -37,6 +37,7 @@ Changelog for NeoFS Node
 - `neofs-adm morph set-config` now supports well-known `MaintenanceModeAllowed` key (#1892)
 - `add`, `get-by-path` and `add-by-path` tree service CLI commands (#1332)
 - Tree synchronisation on startup (#1329)
+- Morph client returns to the highest priority endpoint after the switch (#1615)
 
 ### Changed
 - Allow to evacuate shard data with `EvacuateShard` control RPC (#1800)